edit BackTop: convert to tsx
